+++ test/apple-pay/unit/apple-pay.js
@@ -406,7 +406,7 @@ describe('ApplePay', function () {
           displayName: 'JS SDK Integration'
         }, function () {});
 
-        expect(analytics.sendEvent).to.have.been.calledWith(this.client, 'applepay.performValidation.succeeded');
+        expect(analytics.sendEvent).to.have.been.calledWith(this.client, 'web.applepay.performValidation.succeeded');
       });
 
       it('submits failed', function () {
@@ -421,7 +421,7 @@ describe('ApplePay', function () {
           displayName: 'JS SDK Integration'
         }, function () {});
 
-        expect(analytics.sendEvent).to.have.been.calledWith(this.client, 'applepay.performValidation.failed');
+        expect(analytics.sendEvent).to.have.been.calledWith(this.client, 'web.applepay.performValidation.failed');
       });
     });
 
@@ -439,7 +439,7 @@ describe('ApplePay', function () {
           token: 'token'
         }, function () {});
 
-        expect(analytics.sendEvent).to.have.been.calledWith(this.client, 'applepay.tokenize.succeeded');
+        expect(analytics.sendEvent).to.have.been.calledWith(this.client, 'web.applepay.tokenize.succeeded');
       });
 
       it('submits failed', function () {
@@ -453,7 +453,7 @@ describe('ApplePay', function () {
           token: 'token'
         }, function () {});
 
-        expect(analytics.sendEvent).to.have.been.calledWith(this.client, 'applepay.tokenize.failed');
+        expect(analytics.sendEvent).to.have.been.calledWith(this.client, 'web.applepay.tokenize.failed');
       });
     });
   });